Change preparing the output directory part
Since the `django-admin.py startproject` creates the root of the output directory and raise an error if the root directory already exists, this line (creating the root of the output directory) before running the `django-admin.py startproject` raise errors. Changing it to create the parent directory of the output root prevents the error.